+++ src/plugins/Dashboard/index.js
@@ -4,11 +4,12 @@ const dragDrop = require('drag-drop')
 const DashboardUI = require('./Dashboard')
 const StatusBar = require('../StatusBar')
 const Informer = require('../Informer')
+const ThumbnailGenerator = require('../ThumbnailGenerator')
 const { findAllDOMElements, toArray } = require('../../core/Utils')
 const prettyBytes = require('prettier-bytes')
 const { defaultTabIcon } = require('./icons')
 
-// some code for managing focus was adopted from https://github.com/ghosh/micromodal
+// Some code for managing focus was adopted from https://github.com/ghosh/micromodal
 // MIT licence, https://github.com/ghosh/micromodal/blob/master/LICENSE.md
 // Copyright (c) 2017 Indrashish Ghosh
 const FOCUSABLE_ELEMENTS = [
@@ -71,6 +72,7 @@ module.exports = class Dashboard extends Plugin {
       inline: false,
       width: 750,
       height: 550,
+      thumbnailWidth: 280,
       semiTransparent: false,
       defaultTabIcon: defaultTabIcon,
       showProgressDetails: false,
@@ -78,8 +80,11 @@ module.exports = class Dashboard extends Plugin {
       hideProgressAfterFinish: false,
       note: null,
       closeModalOnClickOutside: false,
-      locale: defaultLocale,
-      onRequestCloseModal: () => this.closeModal()
+      disableStatusBar: false,
+      disableInformer: false,
+      disableThumbnailGenerator: false,
+      onRequestCloseModal: () => this.closeModal(),
+      locale: defaultLocale
     }
 
     // merge default options with the ones set by user
@@ -91,9 +96,9 @@ module.exports = class Dashboard extends Plugin {
     this.translator = new Translator({locale: this.locale})
     this.i18n = this.translator.translate.bind(this.translator)
 
+    this.openModal = this.openModal.bind(this)
     this.closeModal = this.closeModal.bind(this)
     this.requestCloseModal = this.requestCloseModal.bind(this)
-    this.openModal = this.openModal.bind(this)
     this.isModalOpen = this.isModalOpen.bind(this)
 
     this.addTarget = this.addTarget.bind(this)
@@ -202,12 +207,14 @@ module.exports = class Dashboard extends Plugin {
     })
 
     // save scroll position
-    this.savedDocumentScrollPosition = window.scrollY
+    this.savedScrollPosition = window.scrollY
+    // save active element, so we can restore focus when modal is closed
+    this.savedActiveElement = document.activeElement
 
     // add class to body that sets position fixed, move everything back
     // to scroll position
     document.body.classList.add('uppy-Dashboard-isOpen')
-    document.body.style.top = `-${this.savedDocumentScrollPosition}px`
+    document.body.style.top = `-${this.savedScrollPosition}px`
 
     this.updateDashboardElWidth()
     this.setFocusToFirstNode()
@@ -220,7 +227,9 @@ module.exports = class Dashboard extends Plugin {
 
     document.body.classList.remove('uppy-Dashboard-isOpen')
 
-    window.scrollTo(0, this.savedDocumentScrollPosition)
+    this.savedActiveElement.focus()
+
+    window.scrollTo(0, this.savedScrollPosition)
   }
 
   isModalOpen () {
@@ -505,6 +514,12 @@ module.exports = class Dashboard extends Plugin {
       })
     }
 
+    if (!this.opts.disableThumbnailGenerator) {
+      this.uppy.use(ThumbnailGenerator, {
+        thumbnailWidth: this.opts.thumbnailWidth
+      })
+    }
+
     this.discoverProviderPlugins()
 
     this.initEvents()
@@ -513,14 +528,19 @@ module.exports = class Dashboard extends Plugin {
   uninstall () {
     if (!this.opts.disableInformer) {
       const informer = this.uppy.getPlugin('Informer')
+      // Checking if this plugin exists, in case it was removed by uppy-core
+      // before the Dashboard was.
       if (informer) this.uppy.removePlugin(informer)
     }
 
     if (!this.opts.disableStatusBar) {
       const statusBar = this.uppy.getPlugin('StatusBar')
-      // Checking if this plugin exists, in case it was removed by uppy-core
-      // before the Dashboard was.
       if (statusBar) this.uppy.removePlugin(statusBar)
+    }
+
+    if (!this.opts.disableThumbnailGenerator) {
+      const thumbnail = this.uppy.getPlugin('ThumbnailGenerator')
+      if (thumbnail) this.uppy.removePlugin(thumbnail)
     }
 
     const plugins = this.opts.plugins || []